requirements.txt
pytest<=2.9.0,>=2.7.3
pytest-allure-adaptor
selenium